Separate class Player
Separete class player on PlayerWallet. PlayerHealth
