+++ Assets/Scripts/Player/PlayerCollision.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+[RequireComponent(typeof(Player))]
+public class PlayerCollision : MonoBehaviour
+{
+    private Player _player;
+
+    private void Start()
+    {
+        _player = GetComponent<Player>();
+    }
+
+
+}
