--- Assets/Scripts/Player/PlayerCollisionHandler.cs
@@ -1,21 +1,0 @@
-using UnityEngine;
-
-[RequireComponent(typeof(Player))]
-public class PlayerCollisionHandler : MonoBehaviour
-{
-    private Player _player;
-
-    private void Start()
-    {
-        _player = GetComponent<Player>();
-    }
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (other.TryGetComponent(out Coin coin))
-        {
-            _player.AddMoney(coin.Ammount);
-            Destroy(coin.gameObject);
-        }
-    }
-}
