+++ Assets/Scripts/Player/PlayerWallet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3ce048d3024af6439884c248deede9f
+guid: 9801d058bbc97ff4f9531673502bb76b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
