+++ Assets/Scripts/SaveProgress.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7805786921ed12042a370924be03fa33
+guid: 4065367da5369e14b81e995fa002acd9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
