+++ Packages/manifest.json
@@ -1,12 +1,12 @@
 {
   "dependencies": {
     "com.unity.collab-proxy": "1.17.0",
-    "com.unity.ide.rider": "2.0.7",
+    "com.unity.ide.rider": "3.0.7",
     "com.unity.ide.visualstudio": "2.0.16",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.test-framework": "1.1.31",
     "com.unity.textmeshpro": "3.0.6",
-    "com.unity.timeline": "1.4.8",
+    "com.unity.timeline": "1.6.2",
     "com.unity.ugui": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
