+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.3.37f1
-m_EditorVersionWithRevision: 2020.3.37f1 (8c66806a0c04)
+m_EditorVersion: 2021.2.3f1
+m_EditorVersionWithRevision: 2021.2.3f1 (32358a8527b4)
